test(e2e): drive cf-workers file-modifier check via public execute()
Replace the private-method access (`composio.tools.applyBeforeExecuteModifiers`
reached through an `as unknown` cast) with an in-runtime unit test that runs the
real public `execute()` pipeline and stubs only the network tool lookup via
`vi.spyOn(tools, 'getRawComposioToolBySlug')`.

The harness already runs inside workerd through @cloudflare/vitest-pool-workers,
so `vi.spyOn` is available in-runtime and `#file_tool_modifier` still resolves to
the .workerd variant under test. This drops the cast, the InternalToolsForFileModifierTest
structural type, and the contrived /test/file-modifier/error-message worker endpoint
in favor of public, fully type-checked API.
